Zielfunktion kann nun Double-Werte aufnehmen

+++ trunk/simplex4android/src/com/googlecode/simplex4android/SimplexProblemDual.java
@@ -20,7 +20,7 @@ public class SimplexProblemDual extends SimplexProblem {
 	 * @param tableau
 	 * @param target
 	 */
-	public SimplexProblemDual(double[][] tableau, int[] target){ 
+	public SimplexProblemDual(double[][] tableau, double[] target){ 
 		super(tableau, target);
 		SimplexLogic.findPivots(this);
 	    SimplexLogic.calcDeltas(this);
